TCI-17 Implements ProjectService methods to ProjectServiceImpl w/ JPARepository

+++ tci-projects/src/main/java/com/travelci/projects/services/ProjectsService.java
@@ -14,5 +14,5 @@ public interface ProjectsService {
 
     void delete(final ProjectDto projectDto);
 
-    ProjectDto getProjectDetails(final Integer projectId);
+    ProjectDto getProjectDetails(final Long projectId);
 }
